requirements.txt
openenv-core>=0.2.1
pydantic>=2.0
fastapi
uvicorn
python-multipart



